requirements.txt
beautifulsoup4==4.13.4
black==25.1.0
colorama==0.4.6
numpy==2.3.1
pandas==2.3.0
pdfminer.six==20250506
pydantic==2.11.7
PyPDF2==3.0.1
requests==2.32.4
SQLAlchemy==2.0.41
tqdm==4.67.1
